ISSUE-63: merge e ajuste de testes
