+++ backend/src/test/java/br/ufpi/recicle_ai/controller/coleta/ColetaControllerTest.java
@@ -13,6 +13,9 @@ import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -21,6 +24,11 @@ import java.util.Collections;
 
 import static org.hamcrest.Matchers.containsString;
 import static org.mockito.ArgumentMatchers.any;
+import static org.mockito.ArgumentMatchers.anyLong;
+import static org.mockito.ArgumentMatchers.anyString;
+import static org.mockito.ArgumentMatchers.eq;
+import static org.mockito.Mockito.*;
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 
@@ -29,12 +37,11 @@ public class ColetaControllerTest {
 
     @Autowired
     private MockMvc mockMvc;
-
     @Autowired
     private ObjectMapper objectMapper;
-
     @MockBean
     private ColetaService coletaService;
+    private Page<ColetaDTO> mockPage;
 
     @Test
     void createColetas_returnsCreatedAndBody() throws Exception {
@@ -135,4 +142,79 @@ public class ColetaControllerTest {
                         .content(requestJson))
                 .andExpect(status().isBadRequest());
     }
+
+    @Test
+    @DisplayName("GET /coletas - deve retornar todas as coletas com sucesso")
+    void testFindAll() throws Exception {
+        when(coletaService.findAll(any(Pageable.class))).thenReturn(mockPage);
+
+        mockMvc.perform(get("/coletas")
+                        .param("page", "0")
+                        .param("size", "10")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk());
+
+        verify(coletaService, times(1)).findAll(any(Pageable.class));
+    }
+
+    @Test
+    @DisplayName("GET /coletas/por-bairro?bairro=Centro - deve retornar coletas filtradas por bairro")
+    void testFindByBairro() throws Exception {
+        when(coletaService.findByBairro(eq("Centro"), any(Pageable.class))).thenReturn(mockPage);
+
+        mockMvc.perform(get("/coletas/por-bairro")
+                        .param("bairro", "Centro")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk());
+
+        verify(coletaService, times(1)).findByBairro(eq("Centro"), any(Pageable.class));
+    }
+
+    @Test
+    @DisplayName("GET /coletas/por-bairro sem parâmetro - deve retornar erro 400")
+    void testFindByBairroMissingParam() throws Exception {
+        mockMvc.perform(get("/coletas/por-bairro")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isBadRequest());
+
+        verify(coletaService, never()).findByBairro(anyString(), any(Pageable.class));
+    }
+
+    @Test
+    @DisplayName("GET /coletas/coletor/{id} - deve retornar coletas filtradas por coletor")
+    void testFindByColetor() throws Exception {
+        when(coletaService.findByColetor(eq(1L), any(Pageable.class))).thenReturn(mockPage);
+
+        mockMvc.perform(get("/coletas/coletor/{id}", 1L)
+                        .param("page", "0")
+                        .param("size", "10")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk());
+
+        verify(coletaService, times(1)).findByColetor(eq(1L), any(Pageable.class));
+    }
+
+    @Test
+    @DisplayName("GET /coletas/coletor/{id} com id inválido (não numérico) - deve retornar erro 400")
+    void testFindByColetorInvalidId() throws Exception {
+        mockMvc.perform(get("/coletas/coletor/abc")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isBadRequest());
+
+        verify(coletaService, never()).findByColetor(anyLong(), any(Pageable.class));
+    }
+
+    @Test
+    @DisplayName("GET /coletas - deve lidar com serviço retornando página vazia")
+    void testFindAllEmpty() throws Exception {
+        when(coletaService.findAll(any(Pageable.class))).thenReturn(Page.empty(PageRequest.of(0, 10)));
+
+        mockMvc.perform(get("/coletas")
+                        .accept(MediaType.APPLICATION_JSON))
+                .andExpect(status().isOk())
+                .andExpect(jsonPath("$.content").isEmpty());
+
+        verify(coletaService, times(1)).findAll(any(Pageable.class));
+    }
+
 }
